remove remix auth socials dep. rename api dir to server
