+++ app/routes/auth.$provider.callback.tsx
@@ -1,6 +1,6 @@
 import { LoaderFunctionArgs } from "@remix-run/node";
 
-import { authenticator } from "~/api/auth.server";
+import { authenticator } from "~/server/auth.server";
 
 export const loader = ({ request, params }: LoaderFunctionArgs) => {
   console.log("callback", params, request);
