+++ app/routes/projects/repos.tsx
@@ -2,7 +2,7 @@ import { Anchor, Divider, Flex, Text, Title, Tooltip } from "@mantine/core";
 import { useSearchParams } from "@remix-run/react";
 import { IconGitFork, IconUserCircle } from "@tabler/icons-react";
 
-import { getParam, RepoData, sorts, UserScope } from "~/api/projects";
+import { getParam, RepoData, sorts, UserScope } from "~/server/projects";
 import { Card, SortControl } from "~/components";
 import commonStyles from "~/styles/common.module.css";
 import { capitalize, validate } from "~/util";
