+++ app/routes/projects/route.tsx
@@ -1,7 +1,7 @@
 import { LoaderFunctionArgs, MetaFunction, redirect } from "@remix-run/node";
 import { Form, Link, useLoaderData, useSubmit } from "@remix-run/react";
 
-import { getRepos } from "~/api/projects.server";
+import { getRepos } from "~/server/projects.server";
 import { Button } from "~/components";
 import { status } from "~/util";
