+++ app/server/singleton.server.ts
