+++ package.json
@@ -29,7 +29,6 @@
     "octokit": "^3.1.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "remix-auth-socials": "^2.1.0",
     "unstorage": "^1.10.2"
   },
   "devDependencies": {
